add lifts.py, mv Str impl to str.py
