+++ fp/instances/num.py
@@ -41,19 +41,3 @@ class Float(float, metaclass=Alg):
 
     def __str__(self):
         return super().__repr__()
-
-
-class Str(str, metaclass=Monoid):
-    
-    def __str__(self):
-        return super().__repr__() 
-    
-    @classmethod
-    def cast(cls, s):
-        if type(s) is cls:
-            return s
-        if type(s) is str:
-            return cls(s)
-        raise io.CastError(cls, s)
-
-Str.len = Hom(Str, Int)(len)
